backend/mmsds/requirements.txt
pymongo==3.8.0
mysql-connector==2.2.9
pytz==2020.1